@@ -4,12 +4,15 @@ from conans import ConanFile, CMake, tools
 
 class OusterSDKTestConan(ConanFile):
     settings = "os", "compiler", "build_type", "arch"
-    generators = "cmake", "cmake_find_package"
+    generators = "cmake_paths", "cmake_find_package"
 
     def build(self):
         cmake = CMake(self)
         # Current dir is "test_package/build/<build_id>" and CMakeLists.txt is
         # in "test_package"
+        cmake.definitions[
+            "CMAKE_PROJECT_PackageTest_INCLUDE"] = os.path.join(
+                self.build_folder, "conan_paths.cmake")
         cmake.configure()
         cmake.build()
 
@@ -20,5 +23,7 @@ class OusterSDKTestConan(ConanFile):
 
     def test(self):
         if not tools.cross_building(self):
-            os.chdir("bin")
-            self.run(".%sexample_client" % os.sep)
+            os.chdir("examples")
+            # on Windows VS puts execulables under `./Release` or `./Debug` folders
+            exec_path = f"{os.sep}{self.settings.build_type}" if self.settings.os == "Windows" else ""
+            self.run(".%s%sclient_example" % (exec_path, os.sep))
